REF: Prune helper functions and move models to own module
